Improved localization of first quest line
I went through the first set of quests and improved the localization and touched up the grammar in a few places. I tried to keep it as close to the original as I could, but there are a few changes I made that are a bit more significant. Let me know what you think!